Fix test failures on JDK 11.
Tests on JDK 11 because new tests in `neo4j.ogm.metadata` contains a class inherting from `java.lang.Thread`, which is annoted with @Contended. This class is of course found by our „scan the world approach“. Since JDK 9, access to @Contended is prohibited by the module system.

I solved this problem by adapting the tests that scan in `neo4j.ogm.metadata`.

In any case, when a user inadvertently scans a package containing such classes, they will face the same error.
