+++ neo4j-ogm-tests/neo4j-ogm-integration-tests/src/test/java/org/neo4j/ogm/domain/metadata/LabelsAnnotationWithWrongType.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2002-2019 "Neo4j,"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.neo4j.ogm.domain.metadata;
+
+import org.neo4j.ogm.annotation.Labels;
+
+/**
+ * @author Jasper Blues
+ * @author Michael J. Simons
+ */
+public class LabelsAnnotationWithWrongType {
+
+    /**
+     * The labels field must be applied to a type of collection.
+     */
+    @Labels
+    private String labels = "";
+}
