+++ neo4j-ogm-tests/neo4j-ogm-integration-tests/src/test/java/org/neo4j/ogm/domain/metadata/POJO.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.ogm.metadata;
+package org.neo4j.ogm.domain.metadata;
 
 import java.io.Serializable;
 import java.util.List;
