+++ neo4j-ogm-tests/neo4j-ogm-integration-tests/src/test/java/org/neo4j/ogm/metadata/TestMetaDataTypeResolution.java
@@ -26,7 +26,7 @@ import static org.assertj.core.api.Assertions.*;
  */
 public class TestMetaDataTypeResolution {
 
-    private MetaData metaData = new MetaData("org.neo4j.ogm.metadata");
+    private MetaData metaData = new MetaData("org.neo4j.ogm.domain.metadata");
 
     public void checkField(String name, String expectedDescriptor, Class expectedPersistableType) {
         this.checkField(name, expectedDescriptor, expectedPersistableType, null);
